Refactor: extract managers, split monoliths, add structured errors
Phase 1 - Pattern extraction:
- Extract modal backdrop helper (5 modals → shared modal_helper.rs)
- Add IDB readonly transaction helpers (read_one, read_all)
- Deduplicate U8/U16 gate loops in record_decode.rs
- Merge sweep-aware value lookups into base methods

Phase 2 - Structural splits:
- Split timeline.rs (2984 lines → 7 modules under timeline/)
- Decompose RadarGpuRenderer struct (UniformLocations, SweepState, CpuShadowData)
- Split canvas_overlays.rs (1117 lines → 7 modules under canvas_overlays/)

Phase 3 - Manager extractions:
- Extract DownloadQueueManager from main.rs into nexrad/download_queue.rs
- Extract PlaybackManager from main.rs into state/playback_manager.rs

Phase 4 - Cleanup:
- Replace String errors with DataError enum in indexeddb.rs
- Add cached prev-sweep identity to skip redundant resolution
