+++ src/data/indexeddb.rs
@@ -18,6 +18,44 @@ use wasm_bindgen::prelude::*;
 use wasm_bindgen::JsCast;
 use web_sys::{IdbDatabase, IdbObjectStore, IdbRequest, IdbTransaction, IdbTransactionMode};
 
+/// Structured error type for IndexedDB operations.
+#[derive(Debug)]
+#[allow(dead_code)]
+pub enum DataError {
+    /// The database has not been opened yet.
+    NotOpen,
+    /// An IDB transaction failed.
+    TransactionFailed(String),
+    /// An IDB request failed.
+    RequestFailed(String),
+    /// Browser storage quota exceeded.
+    QuotaExceeded { available_mb: f64, required_mb: f64 },
+    /// The requested key was not found.
+    NotFound,
+    /// Deserialization of stored data failed.
+    DeserializationError(String),
+}
+
+impl std::fmt::Display for DataError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            DataError::NotOpen => write!(f, "Database not open"),
+            DataError::TransactionFailed(msg) => write!(f, "Transaction failed: {}", msg),
+            DataError::RequestFailed(msg) => write!(f, "Request failed: {}", msg),
+            DataError::QuotaExceeded {
+                available_mb,
+                required_mb,
+            } => write!(
+                f,
+                "Insufficient storage quota: {:.1} MB available, {:.1} MB required",
+                available_mb, required_mb
+            ),
+            DataError::NotFound => write!(f, "Not found"),
+            DataError::DeserializationError(msg) => write!(f, "Deserialization error: {}", msg),
+        }
+    }
+}
+
 /// Browser storage quota estimate from `navigator.storage.estimate()`.
 #[derive(Debug, Clone, Copy)]
 pub struct StorageQuotaEstimate {
@@ -64,7 +102,7 @@ impl IndexedDbRecordStore {
     }
 
     /// Opens the database, creating/upgrading schema as needed.
-    pub async fn open(&self) -> Result<(), String> {
+    pub async fn open(&self) -> Result<(), DataError> {
         if self.db.borrow().is_some() {
             return Ok(());
         }
@@ -75,7 +113,7 @@ impl IndexedDbRecordStore {
     }
 
     /// Ensures the database is open.
-    async fn ensure_open(&self) -> Result<(), String> {
+    async fn ensure_open(&self) -> Result<(), DataError> {
         if self.db.borrow().is_none() {
             self.open().await?;
         }
@@ -83,11 +121,8 @@ impl IndexedDbRecordStore {
     }
 
     /// Gets the database reference.
-    fn get_db(&self) -> Result<IdbDatabase, String> {
-        self.db
-            .borrow()
-            .clone()
-            .ok_or_else(|| "Database not open".to_string())
+    fn get_db(&self) -> Result<IdbDatabase, DataError> {
+        self.db.borrow().clone().ok_or(DataError::NotOpen)
     }
 
     /// Executes a readwrite transaction on a single object store.
@@ -95,14 +130,14 @@ impl IndexedDbRecordStore {
     /// The closure receives a [`WriteTransaction`] and runs synchronously — no
     /// `.await` is possible inside it, which enforces the IDB rule that
     /// readwrite transactions must not yield to the event loop.
-    async fn write_tx<F, T>(&self, store_name: &str, f: F) -> Result<T, String>
+    async fn write_tx<F, T>(&self, store_name: &str, f: F) -> Result<T, DataError>
     where
-        F: FnOnce(&WriteTransaction) -> Result<T, String>,
+        F: FnOnce(&WriteTransaction) -> Result<T, DataError>,
     {
         let db = self.get_db()?;
         let tx = db
             .transaction_with_str_and_mode(store_name, IdbTransactionMode::Readwrite)
-            .map_err(|e| format!("Failed to create readwrite transaction: {:?}", e))?;
+            .map_err(|e| DataError::TransactionFailed(format!("{:?}", e)))?;
         let result = f(&WriteTransaction::new(&tx))?;
         wait_for_transaction(&tx).await?;
         Ok(result)
@@ -112,9 +147,9 @@ impl IndexedDbRecordStore {
     ///
     /// Same safety guarantee as [`write_tx`]: the closure is synchronous,
     /// preventing any `.await` inside the transaction scope.
-    async fn write_tx_multi<F, T>(&self, store_names: &[&str], f: F) -> Result<T, String>
+    async fn write_tx_multi<F, T>(&self, store_names: &[&str], f: F) -> Result<T, DataError>
     where
-        F: FnOnce(&WriteTransaction) -> Result<T, String>,
+        F: FnOnce(&WriteTransaction) -> Result<T, DataError>,
     {
         let db = self.get_db()?;
         let names = Array::new();
@@ -123,10 +158,42 @@ impl IndexedDbRecordStore {
         }
         let tx = db
             .transaction_with_str_sequence_and_mode(&names, IdbTransactionMode::Readwrite)
-            .map_err(|e| format!("Failed to create readwrite transaction: {:?}", e))?;
+            .map_err(|e| DataError::TransactionFailed(format!("{:?}", e)))?;
         let result = f(&WriteTransaction::new(&tx))?;
         wait_for_transaction(&tx).await?;
         Ok(result)
+    }
+
+    /// Executes a readonly single-key get on an object store.
+    async fn read_one(&self, store_name: &str, key: &str) -> Result<JsValue, DataError> {
+        let db = self.get_db()?;
+        let tx = db
+            .transaction_with_str_and_mode(store_name, IdbTransactionMode::Readonly)
+            .map_err(|e| DataError::TransactionFailed(format!("{:?}", e)))?;
+        let store = tx
+            .object_store(store_name)
+            .map_err(|e| DataError::TransactionFailed(format!("{:?}", e)))?;
+        let request = store
+            .get(&JsValue::from_str(key))
+            .map_err(|e| DataError::RequestFailed(format!("{:?}", e)))?;
+        wait_for_request(&request).await
+    }
+
+    /// Executes a readonly get_all and deserializes results into a Vec.
+    async fn read_all<T: DeserializeOwned>(&self, store_name: &str) -> Result<Vec<T>, DataError> {
+        let db = self.get_db()?;
+        let tx = db
+            .transaction_with_str_and_mode(store_name, IdbTransactionMode::Readonly)
+            .map_err(|e| DataError::TransactionFailed(format!("{:?}", e)))?;
+        let store = tx
+            .object_store(store_name)
+            .map_err(|e| DataError::TransactionFailed(format!("{:?}", e)))?;
+        let request = store
+            .get_all()
+            .map_err(|e| DataError::RequestFailed(format!("{:?}", e)))?;
+        let result = wait_for_request(&request).await?;
+        let array = Array::from(&result);
+        Ok(deserialize_js_array(&array))
     }
 
     // ========================================================================
@@ -142,7 +209,7 @@ impl IndexedDbRecordStore {
     ///
     /// Checks browser storage quota before writing. If remaining quota is
     /// insufficient for the batch, returns an error instead of silently failing.
-    pub async fn put_sweeps_batch(&self, items: &[(String, Vec<u8>)]) -> Result<(), String> {
+    pub async fn put_sweeps_batch(&self, items: &[(String, Vec<u8>)]) -> Result<(), DataError> {
         if items.is_empty() {
             return Ok(());
         }
@@ -154,16 +221,10 @@ impl IndexedDbRecordStore {
             // Require the write size plus 5 MB headroom for IDB overhead/metadata
             let required = batch_size + 5 * 1024 * 1024;
             if remaining < required {
-                return Err(format!(
-                    "Insufficient storage quota: {:.1} MB remaining, need {:.1} MB \
-                     (batch: {:.1} MB). Browser quota: {:.1} MB, usage: {:.1} MB. \
-                     Free browser storage or reduce cache size.",
-                    remaining as f64 / (1024.0 * 1024.0),
-                    required as f64 / (1024.0 * 1024.0),
-                    batch_size as f64 / (1024.0 * 1024.0),
-                    estimate.quota as f64 / (1024.0 * 1024.0),
-                    estimate.usage as f64 / (1024.0 * 1024.0),
-                ));
+                return Err(DataError::QuotaExceeded {
+                    available_mb: remaining as f64 / (1024.0 * 1024.0),
+                    required_mb: required as f64 / (1024.0 * 1024.0),
+                });
             }
         }
 
@@ -176,7 +237,7 @@ impl IndexedDbRecordStore {
                 let buffer = array.buffer();
                 store
                     .put_with_key(&buffer, &JsValue::from_str(key))
-                    .map_err(|e| format!("Failed to put sweep '{}': {:?}", key, e))?;
+                    .map_err(|e| DataError::RequestFailed(format!("{:?}", e)))?;
             }
             Ok(())
         })
@@ -185,23 +246,9 @@ impl IndexedDbRecordStore {
 
     /// Gets a pre-computed sweep blob by key, returning the raw JS ArrayBuffer.
     /// Avoids the 5MB+ copy from JS to Rust that `get_sweep` performs.
-    pub async fn get_sweep_as_js(&self, key: &str) -> Result<Option<ArrayBuffer>, String> {
+    pub async fn get_sweep_as_js(&self, key: &str) -> Result<Option<ArrayBuffer>, DataError> {
         self.ensure_open().await?;
-        let db = self.get_db()?;
-
-        let tx = db
-            .transaction_with_str_and_mode(STORE_SWEEPS, IdbTransactionMode::Readonly)
-            .map_err(|e| format!("Failed to create transaction: {:?}", e))?;
-
-        let store = tx
-            .object_store(STORE_SWEEPS)
-            .map_err(|e| format!("Failed to get sweeps store: {:?}", e))?;
-
-        let request = store
-            .get(&JsValue::from_str(key))
-            .map_err(|e| format!("Failed to get sweep: {:?}", e))?;
-
-        let result = wait_for_request(&request).await?;
+        let result = self.read_one(STORE_SWEEPS, key).await?;
 
         if result.is_undefined() || result.is_null() {
             return Ok(None);
@@ -209,7 +256,7 @@ impl IndexedDbRecordStore {
 
         let buffer: ArrayBuffer = result
             .dyn_into()
-            .map_err(|_| "Expected ArrayBuffer".to_string())?;
+            .map_err(|_| DataError::DeserializationError("Expected ArrayBuffer".to_string()))?;
         Ok(Some(buffer))
     }
 
@@ -218,21 +265,22 @@ impl IndexedDbRecordStore {
     // ========================================================================
 
     /// Writes or updates a scan index entry.
-    pub async fn put_scan_index_entry(&self, entry: &ScanIndexEntry) -> Result<(), String> {
+    pub async fn put_scan_index_entry(&self, entry: &ScanIndexEntry) -> Result<(), DataError> {
         self.ensure_open().await?;
         let storage_key = entry.storage_key();
 
         // Serialize before entering the transaction scope to keep the closure
         // as lean as possible.
-        let json =
-            serde_json::to_string(entry).map_err(|e| format!("Serialization error: {}", e))?;
-        let js = js_sys::JSON::parse(&json).map_err(|e| format!("JSON parse error: {:?}", e))?;
+        let json = serde_json::to_string(entry)
+            .map_err(|e| DataError::DeserializationError(format!("{}", e)))?;
+        let js = js_sys::JSON::parse(&json)
+            .map_err(|e| DataError::DeserializationError(format!("{:?}", e)))?;
 
         self.write_tx(STORE_SCAN_INDEX, |wtx| {
             let store = wtx.object_store(STORE_SCAN_INDEX)?;
             store
                 .put_with_key(&js, &JsValue::from_str(&storage_key))
-                .map_err(|e| format!("Failed to put scan index: {:?}", e))?;
+                .map_err(|e| DataError::RequestFailed(format!("{:?}", e)))?;
             Ok(())
         })
         .await
@@ -242,25 +290,10 @@ impl IndexedDbRecordStore {
     pub async fn scan_availability(
         &self,
         scan: &ScanKey,
-    ) -> Result<Option<ScanIndexEntry>, String> {
+    ) -> Result<Option<ScanIndexEntry>, DataError> {
         self.ensure_open().await?;
-        let db = self.get_db()?;
-
         let storage_key = scan.to_storage_key();
-
-        let tx = db
-            .transaction_with_str_and_mode(STORE_SCAN_INDEX, IdbTransactionMode::Readonly)
-            .map_err(|e| format!("Failed to create transaction: {:?}", e))?;
-
-        let store = tx
-            .object_store(STORE_SCAN_INDEX)
-            .map_err(|e| format!("Failed to get store: {:?}", e))?;
-
-        let request = store
-            .get(&JsValue::from_str(&storage_key))
-            .map_err(|e| format!("Failed to get: {:?}", e))?;
-
-        let result = wait_for_request(&request).await?;
+        let result = self.read_one(STORE_SCAN_INDEX, &storage_key).await?;
         Ok(deserialize_js_value(&result))
     }
 
@@ -270,28 +303,13 @@ impl IndexedDbRecordStore {
         site: &SiteId,
         start: UnixMillis,
         end: UnixMillis,
-    ) -> Result<Vec<ScanIndexEntry>, String> {
+    ) -> Result<Vec<ScanIndexEntry>, DataError> {
         self.ensure_open().await?;
-        let db = self.get_db()?;
+        let entries: Vec<ScanIndexEntry> = self.read_all(STORE_SCAN_INDEX).await?;
 
-        let tx = db
-            .transaction_with_str_and_mode(STORE_SCAN_INDEX, IdbTransactionMode::Readonly)
-            .map_err(|e| format!("Failed to create transaction: {:?}", e))?;
-
-        let store = tx
-            .object_store(STORE_SCAN_INDEX)
-            .map_err(|e| format!("Failed to get store: {:?}", e))?;
-
-        let request = store
-            .get_all()
-            .map_err(|e| format!("Failed to get all: {:?}", e))?;
-
-        let result = wait_for_request(&request).await?;
-        let array = Array::from(&result);
-
-        let mut scans: Vec<ScanIndexEntry> = deserialize_js_array(&array)
+        let mut scans: Vec<ScanIndexEntry> = entries
             .into_iter()
-            .filter(|entry: &ScanIndexEntry| {
+            .filter(|entry| {
                 entry.scan.site.0 == site.0
                     && entry.scan.scan_start >= start
                     && entry.scan.scan_start <= end
@@ -303,53 +321,17 @@ impl IndexedDbRecordStore {
     }
 
     /// Gets total cache size across all scans.
-    pub async fn total_cache_size(&self) -> Result<u64, String> {
+    pub async fn total_cache_size(&self) -> Result<u64, DataError> {
         self.ensure_open().await?;
-        let db = self.get_db()?;
-
-        let tx = db
-            .transaction_with_str_and_mode(STORE_SCAN_INDEX, IdbTransactionMode::Readonly)
-            .map_err(|e| format!("Failed to create transaction: {:?}", e))?;
-
-        let store = tx
-            .object_store(STORE_SCAN_INDEX)
-            .map_err(|e| format!("Failed to get store: {:?}", e))?;
-
-        let request = store
-            .get_all()
-            .map_err(|e| format!("Failed to get all: {:?}", e))?;
-
-        let result = wait_for_request(&request).await?;
-        let array = Array::from(&result);
-
-        let entries: Vec<ScanIndexEntry> = deserialize_js_array(&array);
+        let entries: Vec<ScanIndexEntry> = self.read_all(STORE_SCAN_INDEX).await?;
         let total: u64 = entries.iter().map(|e| e.total_size_bytes).sum();
-
         Ok(total)
     }
 
     /// Gets scans sorted by last_accessed_at (oldest first) for LRU eviction.
-    pub async fn get_lru_scans(&self, limit: u32) -> Result<Vec<ScanIndexEntry>, String> {
+    pub async fn get_lru_scans(&self, limit: u32) -> Result<Vec<ScanIndexEntry>, DataError> {
         self.ensure_open().await?;
-        let db = self.get_db()?;
-
-        let tx = db
-            .transaction_with_str_and_mode(STORE_SCAN_INDEX, IdbTransactionMode::Readonly)
-            .map_err(|e| format!("Failed to create transaction: {:?}", e))?;
-
-        let store = tx
-            .object_store(STORE_SCAN_INDEX)
-            .map_err(|e| format!("Failed to get store: {:?}", e))?;
-
-        let request = store
-            .get_all()
-            .map_err(|e| format!("Failed to get all: {:?}", e))?;
-
-        let result = wait_for_request(&request).await?;
-        let array = Array::from(&result);
-
-        let mut scans: Vec<ScanIndexEntry> = deserialize_js_array(&array);
-
+        let mut scans: Vec<ScanIndexEntry> = self.read_all(STORE_SCAN_INDEX).await?;
         scans.sort_by_key(|s| s.last_accessed_at.0);
         scans.truncate(limit as usize);
         Ok(scans)
@@ -357,7 +339,7 @@ impl IndexedDbRecordStore {
 
     /// Deletes a scan and all its sweep blobs.
     /// Returns the number of bytes freed.
-    pub async fn delete_scan(&self, scan: &ScanKey) -> Result<u64, String> {
+    pub async fn delete_scan(&self, scan: &ScanKey) -> Result<u64, DataError> {
         self.ensure_open().await?;
 
         let scan_storage_key = scan.to_storage_key();
@@ -392,12 +374,12 @@ impl IndexedDbRecordStore {
             for key in &sweep_keys {
                 sweeps_store
                     .delete(&JsValue::from_str(key))
-                    .map_err(|e| format!("Failed to delete sweep: {:?}", e))?;
+                    .map_err(|e| DataError::RequestFailed(format!("{:?}", e)))?;
             }
 
             scan_store
                 .delete(&JsValue::from_str(&scan_storage_key))
-                .map_err(|e| format!("Failed to delete scan index: {:?}", e))?;
+                .map_err(|e| DataError::RequestFailed(format!("{:?}", e)))?;
             Ok(())
         })
         .await?;
@@ -413,7 +395,7 @@ impl IndexedDbRecordStore {
 
     /// Evicts scans until total cache size is below target_bytes.
     /// Returns the number of scans evicted.
-    pub async fn evict_to_size(&self, target_bytes: u64) -> Result<u32, String> {
+    pub async fn evict_to_size(&self, target_bytes: u64) -> Result<u32, DataError> {
         let mut current_size = self.total_cache_size().await?;
         let mut evicted_count = 0u32;
 
@@ -461,23 +443,13 @@ impl IndexedDbRecordStore {
         new_records: u32,
         new_size_bytes: u64,
         new_sweeps: &[SweepMeta],
-    ) -> Result<(), String> {
+    ) -> Result<(), DataError> {
         self.ensure_open().await?;
-        let db = self.get_db()?;
         let storage_key = partial.storage_key();
 
         // --- Readonly tx: read existing entry ---
         let existing: Option<ScanIndexEntry> = {
-            let tx = db
-                .transaction_with_str_and_mode(STORE_SCAN_INDEX, IdbTransactionMode::Readonly)
-                .map_err(|e| format!("Failed to create readonly transaction: {:?}", e))?;
-            let store = tx
-                .object_store(STORE_SCAN_INDEX)
-                .map_err(|e| format!("Failed to get scan_index store: {:?}", e))?;
-            let request = store
-                .get(&JsValue::from_str(&storage_key))
-                .map_err(|e| format!("Failed to get scan index: {:?}", e))?;
-            let result = wait_for_request(&request).await?;
+            let result = self.read_one(STORE_SCAN_INDEX, &storage_key).await?;
             deserialize_js_value(&result)
         };
 
@@ -532,15 +504,16 @@ impl IndexedDbRecordStore {
         };
 
         // --- Readwrite tx: write merged entry ---
-        let json =
-            serde_json::to_string(&merged).map_err(|e| format!("Serialization error: {}", e))?;
-        let js = js_sys::JSON::parse(&json).map_err(|e| format!("JSON parse error: {:?}", e))?;
+        let json = serde_json::to_string(&merged)
+            .map_err(|e| DataError::DeserializationError(format!("{}", e)))?;
+        let js = js_sys::JSON::parse(&json)
+            .map_err(|e| DataError::DeserializationError(format!("{:?}", e)))?;
 
         self.write_tx(STORE_SCAN_INDEX, |wtx| {
             let store = wtx.object_store(STORE_SCAN_INDEX)?;
             store
                 .put_with_key(&js, &JsValue::from_str(&storage_key))
-                .map_err(|e| format!("Failed to put scan index: {:?}", e))?;
+                .map_err(|e| DataError::RequestFailed(format!("{:?}", e)))?;
             Ok(())
         })
         .await
@@ -558,44 +531,30 @@ impl IndexedDbRecordStore {
         archive_start: UnixMillis,
         archive_end_ms: i64,
         exclude_key: &ScanKey,
-    ) -> Result<u32, String> {
+    ) -> Result<u32, DataError> {
         self.ensure_open().await?;
-        let db = self.get_db()?;
 
         // Read all scan index entries
-        let tx = db
-            .transaction_with_str_and_mode(STORE_SCAN_INDEX, IdbTransactionMode::Readonly)
-            .map_err(|e| format!("Failed to create transaction: {:?}", e))?;
-        let store = tx
-            .object_store(STORE_SCAN_INDEX)
-            .map_err(|e| format!("Failed to get store: {:?}", e))?;
-        let request = store
-            .get_all()
-            .map_err(|e| format!("Failed to get all: {:?}", e))?;
-        let result = wait_for_request(&request).await?;
-        let array = Array::from(&result);
+        let all_entries: Vec<ScanIndexEntry> = self.read_all(STORE_SCAN_INDEX).await?;
 
         // Find overlapping scans for this site
         let mut to_delete: Vec<ScanKey> = Vec::new();
-        for i in 0..array.length() {
-            let value = array.get(i);
-            if let Some(entry) = deserialize_js_value::<ScanIndexEntry>(&value) {
-                if entry.scan.site.0 != site.0 {
-                    continue;
-                }
-                if entry.scan == *exclude_key {
-                    continue;
-                }
-                let existing_start = entry.scan.scan_start.0;
-                let existing_end = entry
-                    .end_timestamp_secs
-                    .map(|s| s * 1000)
-                    .unwrap_or(existing_start);
+        for entry in &all_entries {
+            if entry.scan.site.0 != site.0 {
+                continue;
+            }
+            if entry.scan == *exclude_key {
+                continue;
+            }
+            let existing_start = entry.scan.scan_start.0;
+            let existing_end = entry
+                .end_timestamp_secs
+                .map(|s| s * 1000)
+                .unwrap_or(existing_start);
 
-                // Two ranges overlap if start_a <= end_b AND start_b <= end_a
-                if archive_start.0 <= existing_end && existing_start <= archive_end_ms {
-                    to_delete.push(entry.scan.clone());
-                }
+            // Two ranges overlap if start_a <= end_b AND start_b <= end_a
+            if archive_start.0 <= existing_end && existing_start <= archive_end_ms {
+                to_delete.push(entry.scan.clone());
             }
         }
 
@@ -621,7 +580,7 @@ impl IndexedDbRecordStore {
     }
 
     /// Clears all data from all stores.
-    pub async fn clear_all(&self) -> Result<(), String> {
+    pub async fn clear_all(&self) -> Result<(), DataError> {
         // Clear each object store rather than deleting the database.
         // deleteDatabase would hang if any other connection (e.g. the worker)
         // is still open, because the delete is blocked until ALL connections close.
@@ -630,10 +589,10 @@ impl IndexedDbRecordStore {
         self.write_tx_multi(&[STORE_SWEEPS, STORE_SCAN_INDEX], |wtx| {
             wtx.object_store(STORE_SWEEPS)?
                 .clear()
-                .map_err(|e| format!("Failed to clear sweeps store: {:?}", e))?;
+                .map_err(|e| DataError::RequestFailed(format!("{:?}", e)))?;
             wtx.object_store(STORE_SCAN_INDEX)?
                 .clear()
-                .map_err(|e| format!("Failed to clear scan_index store: {:?}", e))?;
+                .map_err(|e| DataError::RequestFailed(format!("{:?}", e)))?;
             Ok(())
         })
         .await?;
@@ -672,10 +631,10 @@ impl<'a> WriteTransaction<'a> {
     }
 
     /// Gets an object store from this transaction.
-    pub fn object_store(&self, name: &str) -> Result<IdbObjectStore, String> {
+    pub fn object_store(&self, name: &str) -> Result<IdbObjectStore, DataError> {
         self.tx
             .object_store(name)
-            .map_err(|e| format!("Failed to get store '{}': {:?}", name, e))
+            .map_err(|e| DataError::TransactionFailed(format!("{:?}", e)))
     }
 }
 
@@ -684,24 +643,28 @@ impl<'a> WriteTransaction<'a> {
 // ============================================================================
 
 /// Gets the IdbFactory from the current global scope (works in both Window and Worker).
-fn get_idb_factory() -> Result<web_sys::IdbFactory, String> {
+fn get_idb_factory() -> Result<web_sys::IdbFactory, DataError> {
     let global = js_sys::global();
     let idb = js_sys::Reflect::get(&global, &wasm_bindgen::JsValue::from_str("indexedDB"))
-        .map_err(|e| format!("Failed to access indexedDB: {:?}", e))?;
+        .map_err(|e| {
+            DataError::TransactionFailed(format!("Failed to access indexedDB: {:?}", e))
+        })?;
     if idb.is_undefined() || idb.is_null() {
-        return Err("IndexedDB not available in this context".to_string());
+        return Err(DataError::TransactionFailed(
+            "IndexedDB not available in this context".to_string(),
+        ));
     }
     idb.dyn_into::<web_sys::IdbFactory>()
-        .map_err(|_| "indexedDB is not an IdbFactory".to_string())
+        .map_err(|_| DataError::TransactionFailed("indexedDB is not an IdbFactory".to_string()))
 }
 
 /// Opens the database, creating schema as needed.
-async fn open_database() -> Result<IdbDatabase, String> {
+async fn open_database() -> Result<IdbDatabase, DataError> {
     let idb_factory = get_idb_factory()?;
 
     let open_request = idb_factory
         .open_with_u32(DATABASE_NAME, DATABASE_VERSION)
-        .map_err(|e| format!("Failed to open database: {:?}", e))?;
+        .map_err(|e| DataError::TransactionFailed(format!("Failed to open database: {:?}", e)))?;
 
     // Set up upgrade handler
     let onupgradeneeded = Closure::wrap(Box::new(move |event: web_sys::IdbVersionChangeEvent| {
@@ -737,7 +700,7 @@ async fn open_database() -> Result<IdbDatabase, String> {
     let db_result = wait_for_request(&open_request).await?;
     let db: IdbDatabase = db_result
         .dyn_into()
-        .map_err(|_| "Failed to cast to IdbDatabase".to_string())?;
+        .map_err(|_| DataError::TransactionFailed("Failed to cast to IdbDatabase".to_string()))?;
 
     log::info!("Opened IndexedDB {} v{}", DATABASE_NAME, DATABASE_VERSION);
 
@@ -745,7 +708,7 @@ async fn open_database() -> Result<IdbDatabase, String> {
 }
 
 /// Waits for an IDB request to complete.
-async fn wait_for_request(request: &IdbRequest) -> Result<JsValue, String> {
+async fn wait_for_request(request: &IdbRequest) -> Result<JsValue, DataError> {
     let (tx, rx) = futures_channel::oneshot::channel::<Result<JsValue, String>>();
     let tx = Rc::new(RefCell::new(Some(tx)));
 
@@ -783,7 +746,9 @@ async fn wait_for_request(request: &IdbRequest) -> Result<JsValue, String> {
     request.set_onsuccess(Some(onsuccess.as_ref().unchecked_ref()));
     request.set_onerror(Some(onerror.as_ref().unchecked_ref()));
 
-    let result = rx.await.map_err(|_| "Channel closed".to_string())?;
+    let result = rx
+        .await
+        .map_err(|_| DataError::RequestFailed("Channel closed".to_string()))?;
 
     request.set_onsuccess(None);
     request.set_onerror(None);
@@ -791,11 +756,11 @@ async fn wait_for_request(request: &IdbRequest) -> Result<JsValue, String> {
     drop(onsuccess);
     drop(onerror);
 
-    result
+    result.map_err(DataError::RequestFailed)
 }
 
 /// Waits for an IDB transaction to complete.
-async fn wait_for_transaction(tx: &IdbTransaction) -> Result<(), String> {
+async fn wait_for_transaction(tx: &IdbTransaction) -> Result<(), DataError> {
     let (sender, rx) = futures_channel::oneshot::channel::<Result<(), String>>();
     let sender = Rc::new(RefCell::new(Some(sender)));
 
@@ -817,7 +782,9 @@ async fn wait_for_transaction(tx: &IdbTransaction) -> Result<(), String> {
     tx.set_oncomplete(Some(oncomplete.as_ref().unchecked_ref()));
     tx.set_onerror(Some(onerror.as_ref().unchecked_ref()));
 
-    let result = rx.await.map_err(|_| "Channel closed".to_string())?;
+    let result = rx
+        .await
+        .map_err(|_| DataError::TransactionFailed("Channel closed".to_string()))?;
 
     tx.set_oncomplete(None);
     tx.set_onerror(None);
@@ -825,7 +792,7 @@ async fn wait_for_transaction(tx: &IdbTransaction) -> Result<(), String> {
     drop(oncomplete);
     drop(onerror);
 
-    result
+    result.map_err(DataError::TransactionFailed)
 }
 
 /// Deserializes a JsValue to a Rust type via JSON.
