+++ src/ui/canvas_overlays/sweep.rs
@@ -1,413 +1,11 @@
-use super::colors::{canvas as canvas_colors, radar, sites as site_colors};
-use crate::data::{get_site, NEXRAD_SITES};
-use crate::geo::{GeoLineRenderer, GlobeCamera, GlobeRenderer, MapProjection};
-use crate::nexrad::{RadarGpuRenderer, RADAR_COVERAGE_RANGE_KM};
-use crate::state::{AppState, GeoLayerVisibility};
-use eframe::egui::{self, Color32, Painter, Pos2, Rect, RichText, Vec2};
-use eframe::egui::{Stroke, StrokeKind};
+use super::super::canvas::{format_age_compact, format_time_short};
+use super::super::colors::{canvas as canvas_colors, radar};
+use crate::geo::MapProjection;
+use crate::nexrad::RADAR_COVERAGE_RANGE_KM;
+use crate::state::AppState;
+use eframe::egui::{self, Color32, Painter, Pos2, Rect, Stroke, Vec2};
 use geo_types::Coord;
-use glow::HasContext;
 use std::f32::consts::PI;
-use std::sync::{Arc, Mutex};
-
-use super::canvas::{
-    age_color, format_age, format_age_compact, format_time_short, format_unix_timestamp_with_date,
-    AGE_RANGE_COLLAPSE_SECS, ARCHIVE_AGE_THRESHOLD_SECS,
-};
-
-#[allow(clippy::too_many_arguments)]
-pub(crate) fn draw_globe(
-    ui: &mut egui::Ui,
-    rect: &Rect,
-    state: &AppState,
-    globe_renderer: Option<&Arc<Mutex<GlobeRenderer>>>,
-    geo_line_renderer: Option<&Arc<Mutex<GeoLineRenderer>>>,
-    gpu_renderer: Option<&Arc<Mutex<RadarGpuRenderer>>>,
-    globe_radar_renderer: Option<&Arc<Mutex<crate::nexrad::GlobeRadarRenderer>>>,
-    volume_ray_renderer: Option<&Arc<Mutex<crate::nexrad::VolumeRayRenderer>>>,
-) {
-    let Some(globe_r) = globe_renderer.cloned() else {
-        return;
-    };
-    let geo_r = geo_line_renderer.cloned();
-    let radar_r = gpu_renderer.cloned();
-    let globe_rr = globe_radar_renderer.cloned();
-    let vol_r = volume_ray_renderer.cloned();
-
-    let camera = state.viz_state.camera.clone();
-    let processing = state.render_processing.clone();
-    let radar_lat = state.viz_state.center_lat;
-    let radar_lon = state.viz_state.center_lon;
-    let volume_enabled = state.viz_state.volume_3d_enabled;
-    let volume_density_cutoff = state.viz_state.volume_density_cutoff;
-    let geo_vis = crate::geo::geo_line_renderer::VisibleLayers {
-        states: state.layer_state.geo.states,
-        counties: state.layer_state.geo.counties,
-        highways: state.layer_state.geo.highways,
-        lakes: state.layer_state.geo.lakes,
-    };
-
-    let callback = egui::PaintCallback {
-        rect: *rect,
-        callback: Arc::new(egui_glow::CallbackFn::new(move |_info, painter| {
-            let gl = painter.gl();
-
-            // 1. Draw globe sphere (sets up depth buffer)
-            if let Ok(gr) = globe_r.lock() {
-                gr.paint(gl, &camera);
-            }
-
-            // 2. Draw geo lines on sphere surface
-            if let Some(ref glr) = geo_r {
-                if let Ok(lr) = glr.lock() {
-                    lr.paint(gl, &camera, &geo_vis);
-                }
-            }
-
-            // 3. Draw radar data on sphere
-            let mut drew_volume = false;
-            if volume_enabled {
-                if let (Some(ref vr), Some(ref rr)) = (&vol_r, &radar_r) {
-                    if let (Ok(mut v), Ok(flat_r)) = (vr.lock(), rr.lock()) {
-                        if v.has_data() {
-                            // Get current viewport dimensions for FBO sizing
-                            let mut vp = [0i32; 4];
-                            unsafe { gl.get_parameter_i32_slice(glow::VIEWPORT, &mut vp) };
-                            v.paint(
-                                gl,
-                                &camera,
-                                flat_r.lut_texture(),
-                                &processing,
-                                flat_r.value_min(),
-                                flat_r.value_range(),
-                                volume_density_cutoff,
-                                vp[2], // viewport width
-                                vp[3], // viewport height
-                            );
-                            drew_volume = true;
-                        }
-                    }
-                }
-            }
-            if !drew_volume {
-                if let (Some(ref rr), Some(ref grr)) = (&radar_r, &globe_rr) {
-                    if let (Ok(flat_r), Ok(mut globe_r)) = (rr.lock(), grr.lock()) {
-                        if flat_r.has_data() {
-                            // Rebuild mesh if site changed
-                            globe_r.update_site(gl, radar_lat, radar_lon, flat_r.max_range_km());
-                            // Paint radar on sphere
-                            globe_r.paint(gl, &camera, &flat_r, &processing);
-                        }
-                    }
-                }
-            }
-
-            // Restore GL state for egui
-            unsafe {
-                gl.disable(glow::DEPTH_TEST);
-                gl.disable(glow::CULL_FACE);
-                gl.depth_mask(false);
-            }
-        })),
-    };
-
-    ui.painter().add(callback);
-}
-
-pub(crate) fn draw_overlay_info(ui: &mut egui::Ui, rect: &Rect, state: &AppState) {
-    let has_prev = state.viz_state.prev_sweep_overlay.is_some();
-    let overlay_pos = rect.left_top() + Vec2::new(10.0, 10.0);
-    let overlay_height = if has_prev { 130.0 } else { 90.0 };
-    let overlay_rect = Rect::from_min_size(overlay_pos, Vec2::new(290.0, overlay_height));
-
-    ui.scope_builder(egui::UiBuilder::new().max_rect(overlay_rect), |ui| {
-        ui.vertical(|ui| {
-            // Show loud "ARCHIVE DATA" banner when data is old enough to be confusable
-            let is_archive = state
-                .viz_state
-                .data_staleness_secs
-                .is_some_and(|s| s > ARCHIVE_AGE_THRESHOLD_SECS);
-            if is_archive {
-                ui.label(
-                    RichText::new("ARCHIVE DATA")
-                        .monospace()
-                        .size(14.0)
-                        .strong()
-                        .color(Color32::from_rgb(255, 160, 40)),
-                );
-            }
-
-            let info_color = Color32::from_rgb(200, 200, 220);
-            let label = if has_prev { "Current" } else { "Site" };
-            ui.label(
-                RichText::new(format!("{}: {}", label, state.viz_state.site_id))
-                    .monospace()
-                    .size(12.0)
-                    .color(info_color),
-            );
-            ui.label(
-                RichText::new(format!("Time: {}", state.viz_state.timestamp))
-                    .monospace()
-                    .size(12.0)
-                    .color(info_color),
-            );
-            ui.label(
-                RichText::new(format!("Elev: {}", state.viz_state.elevation))
-                    .monospace()
-                    .size(12.0)
-                    .color(info_color),
-            );
-            if let Some(end_secs) = state.viz_state.data_staleness_secs {
-                let color = age_color(end_secs);
-                let age_text = if end_secs < AGE_RANGE_COLLAPSE_SECS {
-                    if let Some(start_secs) = state.viz_state.data_staleness_start_secs {
-                        format!("Age: {} – {}", format_age(start_secs), format_age(end_secs),)
-                    } else {
-                        format!("Age: {}", format_age(end_secs))
-                    }
-                } else {
-                    format!("Age: {}", format_age(end_secs))
-                };
-                ui.label(RichText::new(age_text).monospace().size(12.0).color(color));
-            }
-
-            // Previous sweep info during sweep animation
-            if let Some((prev_elev, prev_start, prev_end)) = state.viz_state.prev_sweep_overlay {
-                ui.add_space(2.0);
-                let prev_color = Color32::from_rgb(170, 170, 190);
-                ui.label(
-                    RichText::new("Previous")
-                        .monospace()
-                        .size(12.0)
-                        .color(prev_color),
-                );
-                let prev_time = format_unix_timestamp_with_date(
-                    (prev_start + prev_end) / 2.0,
-                    state.use_local_time,
-                );
-                ui.label(
-                    RichText::new(format!("Time: {}", prev_time))
-                        .monospace()
-                        .size(12.0)
-                        .color(prev_color),
-                );
-                ui.label(
-                    RichText::new(format!("Elev: {:.1}\u{00B0}", prev_elev))
-                        .monospace()
-                        .size(12.0)
-                        .color(prev_color),
-                );
-            }
-        });
-    });
-}
-
-pub(crate) fn draw_compass(ui: &mut egui::Ui, rect: &Rect, camera: &GlobeCamera) {
-    let painter = ui.painter();
-    let radius = 28.0f32;
-    let margin = 16.0f32;
-    let center = Pos2::new(
-        rect.left() + margin + radius,
-        rect.bottom() - margin - radius,
-    );
-
-    // Background circle
-    painter.circle_filled(
-        center,
-        radius + 4.0,
-        Color32::from_rgba_unmultiplied(15, 15, 25, 180),
-    );
-    painter.circle_stroke(
-        center,
-        radius + 4.0,
-        Stroke::new(1.0, Color32::from_rgba_unmultiplied(80, 80, 100, 160)),
-    );
-
-    // Compute compass rotation to match the camera's on-screen orientation.
-    // In SiteOrbit, orbit_bearing is where the camera IS, not where it looks.
-    // The camera looks FROM the bearing TOWARD the site, so the viewing direction
-    // is bearing + 180°. We add π to account for this.
-    let rotation_rad = match camera.mode {
-        crate::geo::camera::CameraMode::SiteOrbit => {
-            std::f32::consts::PI - camera.orbit_bearing.to_radians()
-        }
-        _ => 0.0,
-    } - camera.rotation.to_radians();
-
-    // Cardinal directions
-    let cardinals = [("N", 0.0), ("E", 90.0), ("S", 180.0), ("W", 270.0)];
-    for (label, bearing_deg) in cardinals {
-        let angle = (bearing_deg as f32).to_radians() + rotation_rad;
-        // angle=0 → up (screen -Y), rotating CW
-        let dir = Vec2::new(angle.sin(), -angle.cos());
-        let label_pos = center + dir * (radius - 2.0);
-
-        let (color, size) = if label == "N" {
-            (Color32::from_rgb(255, 80, 80), 13.0)
-        } else {
-            (Color32::from_rgba_unmultiplied(180, 180, 200, 200), 11.0)
-        };
-
-        painter.text(
-            label_pos,
-            egui::Align2::CENTER_CENTER,
-            label,
-            egui::FontId::proportional(size),
-            color,
-        );
-    }
-
-    // Small tick marks for intercardinals
-    for i in 0..8 {
-        let angle = (i as f32 * 45.0).to_radians() + rotation_rad;
-        if i % 2 == 0 {
-            continue; // skip cardinals, already labeled
-        }
-        let dir = Vec2::new(angle.sin(), -angle.cos());
-        let inner = center + dir * (radius - 8.0);
-        let outer = center + dir * (radius - 2.0);
-        painter.line_segment(
-            [inner, outer],
-            Stroke::new(1.0, Color32::from_rgba_unmultiplied(120, 120, 140, 140)),
-        );
-    }
-
-    // Center dot
-    painter.circle_filled(
-        center,
-        2.0,
-        Color32::from_rgba_unmultiplied(150, 150, 170, 160),
-    );
-}
-
-pub(crate) fn draw_color_scale(
-    ui: &mut egui::Ui,
-    rect: &Rect,
-    product: &crate::state::RadarProduct,
-) {
-    use crate::nexrad::color_table::{build_reflectivity_lut, product_value_range};
-    use nexrad_render::Product;
-
-    let product_nr = match product {
-        crate::state::RadarProduct::Reflectivity => Product::Reflectivity,
-        crate::state::RadarProduct::Velocity => Product::Velocity,
-        crate::state::RadarProduct::SpectrumWidth => Product::SpectrumWidth,
-        crate::state::RadarProduct::DifferentialReflectivity => Product::DifferentialReflectivity,
-        crate::state::RadarProduct::CorrelationCoefficient => Product::CorrelationCoefficient,
-        crate::state::RadarProduct::DifferentialPhase => Product::DifferentialPhase,
-        crate::state::RadarProduct::ClutterFilterPower => Product::ClutterFilterPower,
-    };
-
-    let (min_val, max_val) = product_value_range(product_nr);
-
-    // Build the LUT (1024 entries) — for reflectivity uses OKLab, others use crate scale
-    let lut = if matches!(product, crate::state::RadarProduct::Reflectivity) {
-        build_reflectivity_lut(min_val, max_val)
-    } else {
-        let color_scale = crate::nexrad::color_table::continuous_color_scale(product_nr);
-        let lut_size = 1024usize;
-        let mut data = Vec::with_capacity(lut_size * 4);
-        for i in 0..lut_size {
-            let t = i as f32 / (lut_size - 1) as f32;
-            let value = min_val + t * (max_val - min_val);
-            let color = color_scale.color(value);
-            let rgba = color.to_rgba8();
-            data.extend_from_slice(&rgba);
-        }
-        data
-    };
-
-    let bar_width = 16.0f32;
-    let margin = 14.0f32;
-    let top_margin = 20.0f32;
-    let bottom_margin = 20.0f32;
-    let bar_height = (rect.height() - top_margin - bottom_margin).clamp(100.0, 320.0);
-
-    let bar_left = rect.right() - margin - bar_width;
-    let bar_top = rect.top() + top_margin;
-
-    let painter = ui.painter();
-    let lut_size = 1024usize;
-
-    // Draw the color bar as horizontal slices (bottom = low, top = high)
-    let num_slices = bar_height as usize;
-    for s in 0..num_slices {
-        let frac = s as f32 / (num_slices - 1) as f32;
-        let lut_idx = ((1.0 - frac) * (lut_size - 1) as f32) as usize; // flip: top=high
-        let lut_idx = lut_idx.min(lut_size - 1);
-        let r = lut[lut_idx * 4];
-        let g = lut[lut_idx * 4 + 1];
-        let b = lut[lut_idx * 4 + 2];
-        let a = lut[lut_idx * 4 + 3];
-
-        let y = bar_top + s as f32;
-        let slice_rect = Rect::from_min_size(Pos2::new(bar_left, y), Vec2::new(bar_width, 1.5));
-        painter.rect_filled(slice_rect, 0.0, Color32::from_rgba_unmultiplied(r, g, b, a));
-    }
-
-    // Outline
-    let bar_rect = Rect::from_min_size(
-        Pos2::new(bar_left, bar_top),
-        Vec2::new(bar_width, bar_height),
-    );
-    painter.rect_stroke(
-        bar_rect,
-        0.0,
-        Stroke::new(1.0, Color32::from_rgba_unmultiplied(120, 120, 130, 180)),
-        StrokeKind::Outside,
-    );
-
-    // Tick labels
-    let range = max_val - min_val;
-    let tick_step = if range > 200.0 {
-        60.0
-    } else if range > 60.0 {
-        10.0
-    } else if range > 10.0 {
-        5.0
-    } else if range > 2.0 {
-        1.0
-    } else {
-        0.2
-    };
-
-    let label_x = bar_left - 4.0;
-    let mut val = (min_val / tick_step).ceil() * tick_step;
-    while val <= max_val {
-        let frac = (val - min_val) / range;
-        let y = bar_top + bar_height * (1.0 - frac);
-
-        // Tick line
-        painter.line_segment(
-            [Pos2::new(bar_left - 3.0, y), Pos2::new(bar_left, y)],
-            Stroke::new(1.0, Color32::from_rgba_unmultiplied(180, 180, 190, 200)),
-        );
-
-        // Label
-        let label = if tick_step < 1.0 {
-            format!("{:.1}", val)
-        } else {
-            format!("{:.0}", val)
-        };
-        painter.text(
-            Pos2::new(label_x, y),
-            egui::Align2::RIGHT_CENTER,
-            label,
-            egui::FontId::monospace(10.0),
-            Color32::from_rgba_unmultiplied(180, 180, 190, 220),
-        );
-        val += tick_step;
-    }
-
-    // Unit label at top
-    painter.text(
-        Pos2::new(bar_left + bar_width * 0.5, bar_top - 6.0),
-        egui::Align2::CENTER_BOTTOM,
-        product.unit(),
-        egui::FontId::monospace(10.0),
-        Color32::from_rgba_unmultiplied(160, 160, 170, 200),
-    );
-}
 
 pub(crate) fn render_radar_sweep(
     painter: &Painter,
@@ -1047,71 +645,4 @@ fn align_pos(pos: Pos2, size: Vec2, align: egui::Align2) -> Pos2 {
         egui::Align::Max => pos.y - size.y,
     };
     Pos2::new(x, y)
-}
-
-pub(crate) fn render_nexrad_sites(
-    painter: &Painter,
-    projection: &MapProjection,
-    current_site_id: &str,
-    visibility: &GeoLayerVisibility,
-) {
-    let current_site_id_upper = current_site_id.to_uppercase();
-    let (min_lon, min_lat, max_lon, max_lat) = projection.visible_bounds();
-
-    if visibility.nexrad_sites {
-        for site in NEXRAD_SITES.iter() {
-            if site.id == current_site_id_upper {
-                continue;
-            }
-
-            let padding = 2.0;
-            if site.lat < min_lat - padding
-                || site.lat > max_lat + padding
-                || site.lon < min_lon - padding
-                || site.lon > max_lon + padding
-            {
-                continue;
-            }
-
-            let screen_pos = projection.geo_to_screen(Coord {
-                x: site.lon,
-                y: site.lat,
-            });
-
-            painter.circle_filled(screen_pos, 4.0, site_colors::OTHER);
-            painter.circle_stroke(screen_pos, 4.0, Stroke::new(1.0, site_colors::OTHER_STROKE));
-
-            if visibility.labels {
-                painter.text(
-                    screen_pos + Vec2::new(6.0, -2.0),
-                    egui::Align2::LEFT_CENTER,
-                    site.id,
-                    egui::FontId::proportional(10.0),
-                    site_colors::LABEL,
-                );
-            }
-        }
-    }
-
-    if let Some(site) = get_site(&current_site_id_upper) {
-        let screen_pos = projection.geo_to_screen(Coord {
-            x: site.lon,
-            y: site.lat,
-        });
-
-        painter.circle_filled(screen_pos, 6.0, site_colors::CURRENT);
-        painter.circle_stroke(
-            screen_pos,
-            6.0,
-            Stroke::new(1.5, site_colors::CURRENT_STROKE),
-        );
-
-        painter.text(
-            screen_pos + Vec2::new(8.0, -2.0),
-            egui::Align2::LEFT_CENTER,
-            site.id,
-            egui::FontId::proportional(11.0),
-            site_colors::CURRENT_LABEL,
-        );
-    }
 }
