+++ src/ui/event_modal.rs
@@ -221,7 +221,7 @@ pub fn render_event_modal(
         }
     }
 
-    if ctx.input(|i| i.key_pressed(egui::Key::Escape)) {
+    if super::modal_helper::modal_backdrop(ctx, "event_modal_backdrop", 160) {
         state.event_modal_open = false;
         return;
     }
@@ -232,23 +232,6 @@ pub fn render_event_modal(
     } else {
         "Save Event"
     };
-
-    // Semi-transparent backdrop
-    egui::Area::new(egui::Id::new("event_modal_backdrop"))
-        .fixed_pos(egui::Pos2::ZERO)
-        .order(egui::Order::Middle)
-        .show(ctx, |ui| {
-            let screen_rect = ctx.input(|i| i.viewport_rect());
-            let (response, painter) = ui.allocate_painter(screen_rect.size(), egui::Sense::click());
-            painter.rect_filled(
-                screen_rect,
-                0.0,
-                Color32::from_rgba_unmultiplied(0, 0, 0, 160),
-            );
-            if response.clicked() {
-                state.event_modal_open = false;
-            }
-        });
 
     // Modal window
     egui::Window::new(title)
