+++ src/ui/stats_modal.rs
@@ -13,27 +13,10 @@ pub fn render_stats_modal(ctx: &egui::Context, state: &mut AppState) {
         return;
     }
 
-    if ctx.input(|i| i.key_pressed(egui::Key::Escape)) {
+    if super::modal_helper::modal_backdrop(ctx, "stats_modal_backdrop", 160) {
         state.stats_detail_open = false;
         return;
     }
-
-    // Semi-transparent backdrop
-    egui::Area::new(egui::Id::new("stats_modal_backdrop"))
-        .fixed_pos(egui::Pos2::ZERO)
-        .order(egui::Order::Middle)
-        .show(ctx, |ui| {
-            let screen_rect = ctx.input(|i| i.viewport_rect());
-            let (response, painter) = ui.allocate_painter(screen_rect.size(), egui::Sense::click());
-            painter.rect_filled(
-                screen_rect,
-                0.0,
-                Color32::from_rgba_unmultiplied(0, 0, 0, 160),
-            );
-            if response.clicked() {
-                state.stats_detail_open = false;
-            }
-        });
 
     // Modal window
     egui::Window::new("Performance")
